refactor(context): 🤖 move request message preparation
Move provider history slicing and filtering into the prompt assembler module.
