+++ src/node/services/aiService.test.ts
@@ -8,7 +8,10 @@ import * as path from "node:path";
 import { describe, it, expect, beforeEach, afterEach, mock, spyOn } from "bun:test";
 
 import { AIService, resolveMuxProjectRootForHostFs } from "./aiService";
-import { discoverAvailableSubagentsForToolContext } from "./turnContextAssembler";
+import {
+  discoverAvailableSubagentsForToolContext,
+  prepareProviderRequestMessages,
+} from "./turnContextAssembler";
 import {
   normalizeAnthropicBaseURL,
   buildAppAttributionHeaders,
